+++ swing/common-model/src/main/java/is/codion/swing/common/model/component/combobox/DefaultFilterComboBoxModel.java
@@ -384,6 +384,16 @@ class DefaultFilterComboBoxModel<T> implements FilterComboBoxModel<T> {
 			visible.notifyChanges();
 			filtered.notifyChanges();
 		}
+
+		@Override
+		public T itemAt(int index) {
+			return visible.items.get(index);
+		}
+
+		@Override
+		public int indexOf(T item) {
+			return visible.items.indexOf(item);
+		}
 	}
 
 	private final class VisibleItems implements Observable<List<T>> {